Move the table content in one file
